feat(vqvae.py): add optional discriminator to VQVAE model
This commit adds an optional discriminator to the VQVAE model. The discriminator can be enabled by setting the `use_discriminator` parameter to `True` when initializing the model. If the discriminator is enabled, it will be used to calculate the generator loss in addition to the reconstruction loss and commitment loss. If the discriminator is not enabled, only the reconstruction loss and commitment loss will be used to calculate the loss.